+++ src/main/java/com/mdsl/service/CardSchemeService.java
@@ -20,10 +20,14 @@ import com.mdsl.model.dto.response.CardSchemeResponseDto;
 import com.mdsl.model.entity.CardScheme;
 import com.mdsl.model.mapper.CardSchemeMapper;
 import com.mdsl.repository.CardSchemeRepository;
+import com.mdsl.utils.MakerCheckerEngine;
 import com.mdsl.utils.ResponseCode;
 import com.mdsl.utils.enumerations.StatusEnum;
 
+import lombok.RequiredArgsConstructor;
+
 @Service
+@RequiredArgsConstructor
 public class CardSchemeService {
 
 
@@ -32,6 +36,7 @@ public class CardSchemeService {
 
 	@Autowired
 	private CardSchemeMapper cardSchemeMapper;
+    private final MakerCheckerEngine makerCheckerEngine;
 
 	public List<CardSchemeResponseDto> fetchAllCardScheme() {
 
@@ -106,7 +111,9 @@ public class CardSchemeService {
 					cardScheme.setCreatedBy(String.valueOf(userDetails.getId()));
 				}
 			}
-
+			if (makerCheckerEngine.processIfRequired(request, CardSchemeService.class.getName(), "saveOrUpdateCardScheme", "")) {
+				return null;
+			}
 			CardScheme saved = cardSchemeRepository.save(cardScheme);
 			return cardSchemeMapper.toDto(saved);
 
@@ -118,6 +125,9 @@ public class CardSchemeService {
 	public void deleteCardSchemeById(String id) throws Exception {
 		cardSchemeRepository.findById(id)
 				.orElseThrow(() -> new BusinessException(ResponseCode.CFG_CARDSCHEME_NOT_FOUND, HttpStatus.BAD_REQUEST));
+		if (makerCheckerEngine.processIfRequired(id, CardSchemeService.class.getName(), "deleteCardSchemeById", "")) {
+			return;
+		}
 		cardSchemeRepository.deleteById(id);
 	}
 
@@ -125,6 +135,9 @@ public class CardSchemeService {
 		CardScheme cardScheme = cardSchemeRepository.findById(String.valueOf(changeStatusRequestDto.getIdString()))
 				.orElseThrow(() -> new BusinessException(ResponseCode.CFG_CARDSCHEME_NOT_FOUND, HttpStatus.NOT_FOUND));
 		cardScheme.setStatus(changeStatusRequestDto.getStatus().charAt(0));
+		if (makerCheckerEngine.processIfRequired(changeStatusRequestDto, CardSchemeService.class.getName(), "changeStatus", "")) {
+			return null;
+		}
 		cardSchemeRepository.save(cardScheme);
 
 		return "Card scheme status changed successfully";
